Made employee and gameplay

+++ database/AceDB.sql
@@ -19,8 +19,8 @@ CREATE TABLE `gameplays` (
   `employee_id` int unsigned DEFAULT NULL,
   `emulator_name` varchar(225) DEFAULT NULL,
   `date` date DEFAULT NULL,
-  `rake` int DEFAULT NULL,
-  `count` int DEFAULT NULL,
+  `rake` decimal(12, 2) DEFAULT NULL,
+  `count` decimal(12, 2) DEFAULT NULL,
   `created_at` datetime DEFAULT CURRENT_TIMESTAMP,
   `updated_at` datetime DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
   FOREIGN KEY (`employee_id`) REFERENCES `employees` (`id`)
